wip: add shape tests
