+++ libs/ray/ray_destroy.c
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/14 20:57:20 by vwildner          #+#    #+#             */
-/*   Updated: 2022/08/14 21:03:28 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/08/15 16:32:14 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,4 +32,11 @@ void	destroy_intersections_list(t_intersections *xs)
 	}
 	free(xs->intersections);
 	free(xs);
+}
+
+void	destroy_ray(t_ray *r)
+{
+	free(r->direction);
+	free(r->origin);
+	free(r);
 }
